update scripts & configs

+++ gerl/workers/diffusers_model/pipelines/pipeline_stable_diffusion_3.py
@@ -373,7 +373,7 @@ class StableDiffusion3PipelineWithLogProb(StableDiffusion3Pipeline):
                 latents / self.vae.config.scaling_factor
             ) + self.vae.config.shift_factor
 
-            image = self.vae.decode(latents, return_dict=False)[0]
+            image = self.vae.decode(latents.to(self.vae.dtype), return_dict=False)[0]
             image = self.image_processor.postprocess(image, output_type=output_type)
 
         # Offload all models
